Add tolerance to merge method
There were problems when trying to merge some results at the same
temperature. For some (yet unknown) reason, the program did not see
the temperatures as equal (thogh they were) and did not merge the
results as expected.

In order to fix this, I have changed the merge method to compare
only the first 10 decimal places.